# docs/swagger.yaml
definitions:
...
  dto.Action:
    properties:
      body:
        description: 身體部位(1:全身/2:核心/3:手臂/4:背部/5:臀部/6:腿部/7:肩膀/8:胸部)
        example: 8
        type: integer
      category:
        description: 分類(1:重量訓練/2:有氧/3:HIIT/4:徒手訓練/5:其他)
        example: 1
        type: integer
      cover:
        description: 封面
        example: 32as1d5f13e4.png
        type: string
      equipment:
        description: 器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)
        example: 1
        type: integer
      favorite:
        description: 是否收藏(0:否/1:是)
        example: 1
        type: integer
      id:
        description: 動作id
        example: 1
        type: integer
      intro:
        description: 動作介紹
        example: 槓鈴胸推是很多人在健身房都會訓練的動作，是胸大肌強化最常見的訓練動作
        type: string
      name:
        description: 動作名稱
        example: 槓鈴臥推
        type: string
      source:
        description: 動作來源(1:系統動作/2:教練自創動作)
        example: 2
        type: integer
      type:
        description: 紀錄類型(1:重訓/2:時間長度/3:次數/4:次數與時間/5:有氧)
        example: 1
        type: integer
      video:
        description: 動作影片
        example: 11d547we1d4f8e.mp4
        type: string
    type: object
...
  dto.CourseCustom1:
    properties:
      cover:
        description: 課表封面
        example: f43e5715fe.jpg
        type: string
      id:
        description: 課表 id
        example: 10001
        type: integer
      name:
        description: 課表名稱
        example: henry
        type: string
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 1
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
    type: object
...
  dto.Workout:
    properties:
      end_audio:
        description: 結束語音
        example: d2e15qwe42dw.mp3
        type: string
      equipment:
        description: 所需器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)
        example: 2,3,7
        type: string
      id:
        description: 訓練 id
        example: 1
        type: integer
      name:
        description: 訓練名稱
        example: 第一天胸肌訓練
        type: string
      start_audio:
        description: 前導語音
        example: e6d2131w5q.mp3
        type: string
      workout_set_count:
        description: 動作組數
        example: 1
        type: integer
    type: object
...
  dto.WorkoutLog:
    properties:
      course:
        $ref: '#/definitions/dto.CourseCustom1'
        description: 課表
      create_at:
        description: 創建時間
        example: "2021-05-28 11:00:00"
        type: string
      duration:
        description: 訓練時長
        example: 1
        type: integer
      id:
        description: 紀錄id
        example: 1
        type: integer
      intensity:
        description: 訓練強度(0:未指定/1:輕鬆/2:適中/3:稍難/4:很累)
        example: 1
        type: integer
      place:
        description: 地點(0:未指定/1:住家/2:健身房/3:戶外)
        example: 1
        type: integer
      workout:
        $ref: '#/definitions/dto.Workout'
        description: 訓練
      workout_set_logs:
        description: 訓練組
        items:
          $ref: '#/definitions/dto.WorkoutSetLog'
        type: array
    type: object
  dto.WorkoutLogSummary:
    properties:
      create_at:
        description: 創建時間
        example: "2021-05-28 11:00:00"
        type: string
      duration:
        description: 訓練時長
        example: 1
        type: integer
      id:
        description: 紀錄id
        example: 1
        type: integer
      intensity:
        description: 訓練強度(0:未指定/1:輕鬆/2:適中/3:稍難/4:很累)
        example: 1
        type: integer
      place:
        description: 地點(0:未指定/1:住家/2:健身房/3:戶外)
        example: 1
        type: integer
      workout:
        $ref: '#/definitions/dto.Workout'
        description: 訓練
    type: object
  dto.WorkoutSet:
    properties:
      action:
        $ref: '#/definitions/dto.Action'
        description: 動作
      auto_next:
        description: 自動下一組(Y:是/N:否)
        example: "N"
        type: string
      distance:
        description: 距離(公里)
        example: 0
        type: number
      duration:
        description: 時長(秒)
        example: 30
        type: integer
      id:
        description: 訓練組id
        example: 10
        type: integer
      incline:
        description: 坡度
        example: 0
        type: number
      progress_audio:
        description: 進行中語音
        example: 1d2w3e51d3w.mp3
        type: string
      remark:
        description: 備註
        example: ˋ注意姿勢
        type: string
      reps:
        description: 次數
        example: 0
        type: integer
      start_audio:
        description: 前導語音
        example: 1d2w3e51d3w.mp3
        type: string
      type:
        description: 動作類別(1:動作/2:休息)
        example: 2
        type: integer
      weight:
        description: 重量(公斤)
        example: 0
        type: number
    type: object
...
  dto.WorkoutSetLog:
    properties:
      distance:
        description: 距離(公里)
        example: 1
        type: number
      duration:
        description: 時長(秒)
        example: 30
        type: integer
      id:
        description: 訓練組紀錄id
        type: integer
      incline:
        description: 坡度
        example: 5
        type: number
      name:
        description: 訓練名稱
        type: string
      reps:
        description: 次數
        example: 5
        type: integer
      weight:
        description: 重量(公斤)
        example: 10
        type: number
      workout_set:
        $ref: '#/definitions/dto.WorkoutSet'
        description: 訓練組
    type: object
...
  model.Data:
    type: object
  model.ErrorResult:
    properties:
      code:
        description: 錯誤碼
        example: 9000
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 錯誤訊息
        example: system error!
        type: string
    type: object
...
  model.SuccessResult:
    properties:
      code:
        description: 狀態碼
        example: 0
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 成功訊息
        example: success!
        type: string
    type: object
...
info:
  contact: {}
  description: 健身平台 api
  title: fitness api
paths:
...
  /v1/workout_log/{workout_log_id}:
    delete:
      consumes:
      - application/json
      description: 用於歷史頁面刪除訓練紀錄
      parameters:
      - description: 訓練記錄id
        in: path
        name: workout_log_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            $ref: '#/definitions/model.SuccessResult'
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 刪除訓練紀錄
      tags:
      - History_v1
    get:
      consumes:
      - application/json
      description: 獲取訓練紀錄詳細
      parameters:
      - description: 訓練記錄id
        in: path
        name: workout_log_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorkoutLog'
              type: object
        "400":
          description: 失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取訓練紀錄詳細
      tags:
      - History_v1
  /v1/workout_logs:
    get:
      consumes:
      - application/json
      description: 以日期區間獲取訓練記錄，用於獲取歷史首頁資料
      parameters:
      - description: 區間開始日期 YYYY-MM-DD
        in: query
        name: start_date
        required: true
        type: string
      - description: 區間結束日期 YYYY-MM-DD
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.WorkoutLogSummary'
                  type: array
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 以日期區間獲取訓練記錄
      tags:
      - History_v1
  /v1/workout_product/{workout_id}/workout_sets:
    get:
      consumes:
      - application/json
      description: 獲取訓練組列表(探索區課表)
      parameters:
      - description: 訓練id
        in: path
        name: workout_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.WorkoutSet'
                  type: array
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取訓練組列表(探索區課表)
      tags:
      - Explore_v1
...
securityDefinitions:
  fitness_token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"